Refactored visualizations data generation code
